+++ molo/polls/wagtail_hooks.py
@@ -2,7 +2,8 @@ from django.conf.urls import url
 from molo.polls.admin import QuestionsModelAdmin
 from molo.polls.admin_views import QuestionResultsAdminView
 from wagtail.wagtailcore import hooks
-from wagtailmodeladmin.options import wagtailmodeladmin_register
+from wagtail.contrib.modeladmin.options import modeladmin_register
+from django.contrib.auth.models import User
 
 
 @hooks.register('register_admin_urls')
@@ -13,4 +14,13 @@ def register_question_results_admin_view_url():
             name='question-results-admin'),
     ]
 
-wagtailmodeladmin_register(QuestionsModelAdmin)
+
+modeladmin_register(QuestionsModelAdmin)
+
+
+@hooks.register('construct_main_menu')
+def show_polls_entries_for_users_have_access(request, menu_items):
+    if not request.user.is_superuser and not User.objects.filter(
+            pk=request.user.pk, groups__name='Moderators').exists():
+        menu_items[:] = [
+            item for item in menu_items if item.name != 'polls']