+++ CHANGES.rst
@@ -1,6 +1,29 @@
 CHANGE LOG
 ==========
 
+3.0.0
+-----
+- Removed dependency on wagtailmodeladmin
+
+Backwards incompatible changes
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+- Removed use of ``wagtailmodeladmin``: use ``wagtail.contrib.modeladmin`` instead
+- ``{% load wagtailmodeladmin_tags %}`` has been replaced by ``{% load modeladmin_tags %}``
+
+NOTE: This release is not compatible with molo versions that are less than 4.0
+
+2.2.1
+-----
+- Add polls permissions to groups
+
+2.2.1
+-----
+- Return None if there is no question
+
+2.2.0
+-----
+- Add support for hiding untranslated content
+
 2.1.0
 -----
 - Added Polls view to Wagtail Admin